// frontend/src/api/models/BasketUpdatePageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ShortVariantProduct } from './ShortVariantProduct';
export type BasketUpdatePageResponse = {
	readonly id: string;
	quantity?: number;
	readonly variant: ShortVariantProduct;
};